Support Typescript version numbers

+++ scripts/scanTypescript.sh
@@ -4,19 +4,30 @@
 
 # Uses: patchJQAssistantTypescriptPlugin.sh, detectChangedFiles.sh
 
+# Note: This script must not output anything except for the return code to stdout.
+#       All output of the scanning needs to be redirected to stderr using ">&2".
+
 # Fail on any error ("-e" = exit on first error, "-o pipefail" exist on errors within piped commands)
 set -o errexit -o pipefail
 
 ARTIFACTS_DIRECTORY=${ARTIFACTS_DIRECTORY:-"artifacts"}
 SOURCE_DIRECTORY=${SOURCE_DIRECTORY:-"source"}
+
 TYPESCRIPT_SCAN_HEAP_MEMORY=${TYPESCRIPT_SCAN_HEAP_MEMORY:-"4096"} # Heap memory in megabytes for Typescript scanning with (Node.js process). Defaults to 4096 MB.
+echo "scanTypescript: TYPESCRIPT_SCAN_HEAP_MEMORY=${TYPESCRIPT_SCAN_HEAP_MEMORY}" >&2
 
 ## Get this "scripts" directory if not already set
 # Even if $BASH_SOURCE is made for Bourne-like shells it is also supported by others and therefore here the preferred solution. 
 # CDPATH reduces the scope of the cd command to potentially prevent unintended directory changes.
 # This way non-standard tools like readlink aren't needed.
 SCRIPTS_DIR=${SCRIPTS_DIR:-$( CDPATH=. cd -- "$(dirname -- "${BASH_SOURCE[0]}")" && pwd -P )} # Repository directory containing the shell scripts
-echo "scanTypescript SCRIPTS_DIR=${SCRIPTS_DIR}" >&2
+echo "scanTypescript: SCRIPTS_DIR=${SCRIPTS_DIR}" >&2
+
+# Dry run for internal testing (not intended to be accessible from the outside)
+TYPESCRIPT_SCAN_DRY_RUN=false
+if [ "${TYPESCRIPT_SCAN_DRY_RUN}" = true ] ; then
+    echo "scanTypescript: -> DRY RUN <- Scanning will only be logged, not executed." >&2
+fi
 
 if [ ! -d "./${SOURCE_DIRECTORY}" ] ; then
     echo "scanTypescript: Source directory '${SOURCE_DIRECTORY}' doesn't exist. The scan will therefore be skipped." >&2
@@ -32,9 +43,11 @@ fi
 changeDetectionHashFilePath="./${SOURCE_DIRECTORY}/typescriptFileChangeDetectionHashFile.txt"
 changeDetectionReturnCode=$( source "${SCRIPTS_DIR}/detectChangedFiles.sh" --readonly --hashfile "${changeDetectionHashFilePath}" --paths "./${SOURCE_DIRECTORY}")
 
-if [ "${changeDetectionReturnCode}" == "0" ] ; then
+if [ "${changeDetectionReturnCode}" == "0" ]; then
     echo "scanTypescript: Files unchanged. Scan skipped."
-else
+fi
+
+if [ "${changeDetectionReturnCode}" != "0" ] || [ "${TYPESCRIPT_SCAN_DRY_RUN}" = true ]; then
     echo "scanTypescript: Detected change (${changeDetectionReturnCode}). Scanning Typescript source using @jqassistant/ts-lce."
     
     mkdir -p "./runtime/logs"
@@ -42,6 +55,7 @@ else
     echo "scanTypescript: LOG_DIRECTORY=${LOG_DIRECTORY}" >&2
 
     source_directories=$( find -L "./${SOURCE_DIRECTORY}" -mindepth 1 -maxdepth 1 -type d  -print0 | xargs -0 -r -I {} echo {} )
+    
     total_directories=$(echo "${source_directories}" | wc -l | awk '{print $1}')
     processed_directories=0
 
@@ -50,13 +64,16 @@ else
         processed_directories=$((processed_directories + 1))
         echo "" >&2
         echo "scanTypescript: $(date +'%Y-%m-%dT%H:%M:%S%z') Scanning ${directory_name} (${processed_directories}/${total_directories}) -----------------" >&2
-        # Note: This script must not output anything except for the return code to stdout,
-        #       all output of the scanning needs to be redirected to stderr using ">&2".
-        #       For later troubleshooting, the output is also copied to a dedicated log file using "tee".
-        # Note: Don't worry about the hardcoded version number. It will be updated by Renovate using a custom Manager.
-        # Note: NODE_OPTIONS --max-old-space-size=4096 increases the memory for larger projects to scan
-        NODE_OPTIONS="${NODE_OPTIONS} --max-old-space-size=${TYPESCRIPT_SCAN_HEAP_MEMORY}" npx --yes @jqassistant/ts-lce@1.3.0 "${directory}" --extension React 2>&1 | tee "${LOG_DIRECTORY}/jqassistant-typescript-scan-${directory_name}.log" >&2
+        
+        if [ "${TYPESCRIPT_SCAN_DRY_RUN}" = false ] ; then
+            # Note: For later troubleshooting, the output is also copied to a dedicated log file using "tee".
+            # Note: Don't worry about the hardcoded version number. It will be updated by Renovate using a custom Manager.
+            # Note: NODE_OPTIONS --max-old-space-size=4096 increases the memory for scanning larger projects
+            NODE_OPTIONS="${NODE_OPTIONS} --max-old-space-size=${TYPESCRIPT_SCAN_HEAP_MEMORY}" npx --yes @jqassistant/ts-lce@1.3.0 "${directory}" --extension React 2>&1 | tee "${LOG_DIRECTORY}/jqassistant-typescript-scan-${directory_name}.log" >&2
+        fi
     done
 
-    changeDetectionReturnCode=$( source "${SCRIPTS_DIR}/detectChangedFiles.sh" --hashfile "${changeDetectionHashFilePath}" --paths "./${SOURCE_DIRECTORY}")
+    if [ "${TYPESCRIPT_SCAN_DRY_RUN}" = false ] ; then
+        changeDetectionReturnCode=$( source "${SCRIPTS_DIR}/detectChangedFiles.sh" --hashfile "${changeDetectionHashFilePath}" --paths "./${SOURCE_DIRECTORY}")
+    fi
 fi
